refactor: Dialog
